Move and merge CreateRoom
